chore: move sync_or_fork and related code to gh.ts
